% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageage.R
\name{ageage}
\alias{ageage}
\title{age error model}
\usage{
ageage(
  read_tester_loc_file,
  species,
  admb_home = NULL,
  recage = 2,
  plus_age = 45,
  max_age = 100,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
age error model
}
