% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_catch.R
\name{clean_catch}
\alias{clean_catch}
\title{clean up catch data}
\usage{
clean_catch(year, TAC, old_catch = NULL)
}
\arguments{
\item{old_catch}{\itemize{
\item if catch is frozen place the file in user_input folder
}}
}
\value{

}
\description{
clean up catch data
}
