% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw.R
\name{clean_raw}
\alias{clean_raw}
\title{clean up all the raw data}
\usage{
clean_raw(
  year,
  species,
  TAC,
  admb_home,
  rec_age,
  plus_age,
  reader_tester = NULL
)
}
\arguments{
\item{year}{model year}

\item{species}{e.g., "NORK"}

\item{TAC}{past three years TAC c(year-3, year-2, year-1)}

\item{admb_home}{location of admb on your computer e.g., "C:/Program Files (x86)/ADMB-12.1"}

\item{rec_age}{recruitment age}

\item{plus_age}{plus age group}

\item{reader_tester}{if there is an alternate reader_tester (age error) file}
}
\value{

}
\description{
clean up all the raw data
}
