% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_row.R
\name{collapse_row}
\alias{collapse_row}
\title{helper function for creating dat file}
\usage{
collapse_row(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
helper function for creating dat file
}
