% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_dat.R
\name{concat_dat}
\alias{concat_dat}
\title{Concatenate a .dat file}
\usage{
concat_dat(
  year,
  model,
  rec_age,
  plus_age,
  spawn_mo = 5,
  n_ageage = NULL,
  n_sizeage = NULL
)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder that the .tpl will be in}

\item{rec_age}{recruitment age}

\item{plus_age}{plus age group}

\item{spawn_mo}{spawning month}

\item{n_sizeage}{of size at age transmission matrices}

\item{n_ageagenumber}{of age error transmission matrices}
}
\value{

}
\description{
Concatenate a .dat file
}
