% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_table.R
\name{fac_table}
\alias{fac_table}
\title{fishery age comp table}
\usage{
fac_table(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}
}
\value{

}
\description{
fishery age comp table
}
\examples{
fac_table(year, model)
}
