% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_age_comp.R
\name{fish_age_comp}
\alias{fish_age_comp}
\title{Fishery age composition}
\usage{
fish_age_comp(year, rec_age, plus_age)
}
\arguments{
\item{plus_age}{}
}
\value{

}
\description{
Fishery age composition
}
