% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_size_comp.R
\name{fish_size_comp}
\alias{fish_size_comp}
\title{Fishery size composition data}
\usage{
fish_size_comp(year, recage, lenbins = NULL)
}
\arguments{
\item{recage}{}
}
\value{

}
\description{
Fishery size composition data
}
