% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc_table.R
\name{fsc_table}
\alias{fsc_table}
\title{fishery size comp table}
\usage{
fsc_table(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}
}
\value{

}
\description{
fishery size comp table
}
\examples{
fsc_table(year, model)
}
