% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_table.R
\name{like_table}
\alias{like_table}
\title{Likelihood table to generate table 11 in the SAFE}
\usage{
like_table(year, model, model_name)
}
\arguments{
\item{year}{model assessment year}

\item{model}{the folder the model is in e.g., m18.2}

\item{model_name}{name of the tpl file (don't include the .tpl extension)}
}
\value{

}
\description{
Likelihood table to generate table 11 in the SAFE
}
\examples{
like_table(year = 2020, model = "m18.2", model_name = "updated_nr")
}
