% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldir.R
\name{modeldir}
\alias{modeldir}
\title{model function directory and tpl files}
\usage{
modeldir(year)
}
\arguments{
\item{year}{assessment year}
}
\value{

}
\description{
creates folders for storing data and output with the top folder being the assessment year
}
\examples{
modeldir(2020)

}
