% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohn.R
\name{mohn}
\alias{mohn}
\title{Calculate Mohn's rho}
\usage{
mohn(x)
}
\arguments{
\item{x}{retro data - a matrix or dataframe with years as rownames}

\item{peels}{defaults to 10, or NULL to use all retro data}

\item{details}{whether to return the intermediate calculations of relative bias.}

\item{plot}{whether to plot the retrospective trajectories.}
}
\value{

}
\description{
Calculate Mohn's rho
}
\examples{
mohn(data)
}
