% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_table.R
\name{param_table}
\alias{param_table}
\title{param_table to generate table 15 in the SAFE}
\usage{
param_table(year, model, model_name)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}

\item{model_name}{name of the tpl file}
}
\value{

}
\description{
param_table to generate table 15 in the SAFE
}
\examples{
param_table(year = 2020, model = "m18.2b", model_name = "updated_nr")
}
