% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot MCMC estimated biomass}
\usage{
plot_biomass(year, model)
}
\arguments{
\item{year}{model year}

\item{model}{folder the model is in}
}
\value{

}
\description{
Plot MCMC estimated biomass
}
