% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comps.R
\name{plot_comps}
\alias{plot_comps}
\title{Plot age and length compositions}
\usage{
plot_comps(year, model)
}
\arguments{
\item{year}{= assessment year}

\item{model}{= folder the model lives in}
}
\value{

}
\description{
Plot age and length compositions
}
\examples{
plot_comps(year, model)
}
