% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params.R
\name{plot_params}
\alias{plot_params}
\title{Plot parameter histograms}
\usage{
plot_params(year, model, model_name)
}
\arguments{
\item{year}{model year}

\item{model}{foler the model is in}

\item{model_name}{e.g., goa_nr_2020}
}
\value{

}
\description{
Plot parameter histograms
}
