% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase.R
\name{plot_phase}
\alias{plot_phase}
\title{Phase plot}
\usage{
plot_phase(year, model, model_name)
}
\arguments{
\item{year}{model year}

\item{model}{folder model is in}

\item{model_name}{e.g. goa_nr_2020}
}
\value{

}
\description{
Phase plot
}
