% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rec.R
\name{plot_rec}
\alias{plot_rec}
\title{Plot recruitment estimates}
\usage{
plot_rec(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}
}
\value{

}
\description{
Plot recruitment estimates
}
