% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survey.R
\name{plot_survey}
\alias{plot_survey}
\title{Plot survey biomass}
\usage{
plot_survey(year, model)
}
\arguments{
\item{year}{model year}

\item{model}{folder model is in}
}
\value{

}
\description{
Plot survey biomass
}
\examples{
plot_survey(year, model)
}
