% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_swath.R
\name{plot_swath}
\alias{plot_swath}
\title{Swath plot}
\usage{
plot_swath(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{the folder with the model in it}
}
\value{

}
\description{
Swath plot
}
\examples{
plot_swath(year, model)
}
