% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_results}
\alias{process_results}
\title{Process model results for tables and figs}
\usage{
process_results(
  year,
  model,
  model_name,
  rec_age,
  plus_age,
  lenbins = NULL,
  ...
)
}
\arguments{
\item{year}{assessment year}

\item{model}{model being evaluated (folder name)}

\item{model_name}{name of the model e.g., goa_nr_2020}
}
\value{

}
\description{
Process model results for tables and figs
}
