% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_data.R
\name{proj_data}
\alias{proj_data}
\title{project data for "Best F"}
\usage{
proj_data(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}
}
\value{

}
\description{
project data for "Best F"
}
