% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrit.R
\name{purrit}
\alias{purrit}
\title{helper function for comp data}
\usage{
purrit(obs, pred = NULL, rec_age, plus_age, comp = "length", lenbins = NULL)
}
\arguments{
\item{obs}{observed data from .rep file}

\item{pred}{predicted data from .rep file (if used)}

\item{rec_age}{recruitement age}

\item{plus_age}{plus age group}

\item{comp}{\code{age} or \code{length} - default is length}

\item{lenbins}{set to base unless using alt in which case the file should be in the \verb{user_input`` folder and the name needs to be provided e.g., }lengthbins.csv\verb{- the column must be named}len_bin`}
}
\value{

}
\description{
helper function for comp data
}
\examples{
purrit(obs, pred = NULL, rec_age, plus_age, comp = "length", lenbins = "lengthbins.csv")
}
