% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data}
\alias{raw_data}
\title{Raw data call from AFSC and AKFIN servers}
\usage{
raw_data(species, year, afsc_user, afsc_pwd, akfin_user, akfin_pwd, ...)
}
\arguments{
\item{species}{is the species of interest e.g., "NORK", "DUSK"}

\item{year}{is the year you would like data through}

\item{afsc_user}{AFSC server username}

\item{afsc_pwd}{AFSC server password}

\item{akfin_user}{AKFIN server username}

\item{akfin_pwd}{AKFIN server password}

\item{...}{}
}
\value{

}
\description{
Raw data call from AFSC and AKFIN servers
}
\examples{
raw_data("NORK", 2020, afsc_user, afsc_pwd, akfin_user, akfin_pwd)

}
