% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_table.R
\name{recruit_table}
\alias{recruit_table}
\title{recruit table to generate table 14 in the SAFE}
\usage{
recruit_table(year, model, model_name, rec_age)
}
\arguments{
\item{year}{model year}

\item{model}{folder the model is in}

\item{model_name}{name of tpl file}

\item{rec_age}{recruitment age}
}
\value{

}
\description{
recruit table to generate table 14 in the SAFE
}
\examples{
recruit_table(year, model, model_name, rec_age)
}
