% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac_table.R
\name{sac_table}
\alias{sac_table}
\title{survey age comp table}
\usage{
sac_table(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}
}
\value{

}
\description{
survey age comp table
}
\examples{
sac_table(year, model)
}
