% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_table.R
\name{ssc_table}
\alias{ssc_table}
\title{survey size comp table}
\usage{
ssc_table(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}
}
\value{

}
\description{
survey size comp table
}
\examples{
ssc_table(year, model)
}
