% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_age_comp.R
\name{survey_age_comp}
\alias{survey_age_comp}
\title{Survey age comp}
\usage{
survey_age_comp(year, rec_age, plus_age)
}
\arguments{
\item{year}{assessment year}

\item{rec_age}{recruitment age}

\item{plus_age}{plus group age}
}
\value{

}
\description{
Survey age comp
}
\examples{
survey_age_comp(year = 2020, rec_age = 2, plus_age = 45)
}
