% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_biomass.R
\name{survey_biomass}
\alias{survey_biomass}
\title{Get the survey biomass}
\usage{
survey_biomass(year, file = NULL)
}
\arguments{
\item{year}{of interest}

\item{type:}{if not using the design-based abundance, the file name must be stated (e.g. "GAP_VAST.csv")}
}
\value{

}
\description{
Get the survey biomass
}
\examples{

}
