% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_size_comp.R
\name{survey_size_comp}
\alias{survey_size_comp}
\title{Survey size composition}
\usage{
survey_size_comp(year, lenbins = NULL)
}
\arguments{
\item{year}{= the year of assessment}

\item{lenbins}{= length bin file}
}
\value{

}
\description{
Survey size composition
}
