% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_length.R
\name{weight_length}
\alias{weight_length}
\title{compute length at age and length sd}
\usage{
weight_length(year, admb_home, rec_age, plus_age)
}
\arguments{
\item{year=}{analysis year}
}
\value{

}
\description{
compute length at age and length sd
}
