% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_dat.R
\name{concat_dat}
\alias{concat_dat}
\title{Concatenate a .dat file}
\usage{
concat_dat(
  year,
  model,
  rec_age,
  plus_age,
  spawn_mo = 5,
  n_ageage = NULL,
  n_sizeage = NULL,
  lenbins = NULL,
  retro = NULL
)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder that the \code{.tpl} will be in}

\item{rec_age}{recruitment age}

\item{plus_age}{plus age group}

\item{spawn_mo}{spawning month}

\item{n_sizeage}{number of size at age transmission matrices default is 1}

\item{lenbins}{set to base unless using alt in which case the file should be in the \verb{user_input`` folder and the name needs to be provided e.g., }lengthbins.csv\verb{- the column must be named}len_bin`}

\item{n_ageagenumber}{of age error transmission matrices default is 1}
}
\value{

}
\description{
Concatenate a .dat file
}
\examples{
concat_dat(year = 2020, model = "base", rec_age = 2, plus_age = 45)

}
