% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_size_comp.R
\name{fish_size_comp}
\alias{fish_size_comp}
\title{Fishery size composition data}
\usage{
fish_size_comp(YEAR, rec_age, lenbins = NULL)
}
\arguments{
\item{rec_age}{recruitment age}

\item{lenbins}{lenbin file if left NULL it looks for here::here(year, "data", "user_input", "len_bin_labels.csv")}

\item{year}{assessment year}
}
\value{

}
\description{
Fishery size composition data
}
