% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrit.R
\name{purrit}
\alias{purrit}
\title{helper function for comp data}
\usage{
purrit(obs, pred = NULL, rec_age, plus_age, comp = "length", lenbins = NULL)
}
\arguments{
\item{obs}{observed data from .rep file}

\item{pred}{predicted data from .rep file (if used)}

\item{rec_age}{recruitement age}

\item{plus_age}{plus age group}

\item{comp}{"age" or "length"}

\item{lenbins}{set to base unless using alt in which case the location and file name need to be provided}
}
\value{

}
\description{
helper function for comp data
}
